Split broker registrar from workload runner
